requirements.txt
# Core agent dependencies
feedparser==6.0.11
requests==2.32.3
web3==7.14.1
python-dotenv==1.1.0
pandas==2.2.3
numpy==2.2.4

# Streamlit dashboard
streamlit==1.45.1
altair==5.5.0
watchdog==6.0.0

# LangGraph / LangChain pipeline
langgraph==0.4.1
langchain-google-genai==2.1.4
langchain-core==0.3.62

# Sentiment analysis
vaderSentiment==3.3.2

# News and trends data
gnews==0.3.7
pytrends==4.9.2
beautifulsoup4==4.13.3

# MCP integration
mcp==1.7.1
httpx-sse==0.4.0

# Testing
pytest==9.0.2
